chore(mvt): Deduplicate vector tile code
