+++ modules/mvt/src/lib/vector-tile/vector-tile-feature.ts
@@ -4,8 +4,9 @@
 
 // This code is forked from https://github.com/mapbox/vector-tile-js under BSD 3-clause license.
 import {Feature} from '@loaders.gl/schema';
+import {FlatFeature, FlatIndexedGeometry, GeojsonGeometryInfo} from '@loaders.gl/schema';
 import Protobuf from 'pbf';
-import {classifyRings} from '../../helpers/geometry-utils';
+import {classifyRings, classifyRingsFlat, projectToLngLatFlat} from '../../helpers/geometry-utils';
 
 export class VectorTileFeature {
   properties: {[x: string]: string | number | boolean | null};
@@ -16,16 +17,18 @@ export class VectorTileFeature {
   _geometry: number;
   _keys: string[];
   _values: (string | number | boolean | null)[];
-  static get types() {
-    return ['Unknown', 'Point', 'LineString', 'Polygon'];
-  }
+  _geometryInfo: GeojsonGeometryInfo;
 
+  static types: Readonly<string[]> = ['Unknown', 'Point', 'LineString', 'Polygon'];
+
+  // eslint-disable-next-line max-params
   constructor(
     pbf: Protobuf,
     end: number,
     extent: any,
     keys: string[],
-    values: (string | number | boolean | null)[]
+    values: (string | number | boolean | null)[],
+    geometryInfo?: GeojsonGeometryInfo
   ) {
     // Public
     this.properties = {};
@@ -39,8 +42,171 @@ export class VectorTileFeature {
     this._keys = keys;
     this._values = values;
 
+    // Only used by binary tiles
+    this._geometryInfo = geometryInfo!;
+
     pbf.readFields(readFeature, this, end);
   }
+
+  toGeoJSON(
+    options: {x: number; y: number; z: number} | ((data: number[], feature: {extent: any}) => void)
+  ): Feature {
+    const coords = this.loadGeometry();
+
+    if (typeof options === 'function') {
+      return _toGeoJSON(this, coords, options);
+    }
+    const {x, y, z} = options;
+    const size = this.extent * Math.pow(2, z);
+    const x0 = this.extent * x;
+    const y0 = this.extent * y;
+
+    function project(line: number[]) {
+      for (let j = 0; j < line.length; j++) {
+        const p = line[j];
+        p[0] = ((p[0] + x0) * 360) / size - 180;
+        const y2 = 180 - ((p[1] + y0) * 360) / size;
+        p[1] = (360 / Math.PI) * Math.atan(Math.exp((y2 * Math.PI) / 180)) - 90;
+      }
+    }
+    return _toGeoJSON(this, coords, project);
+  }
+
+  /**
+   *
+   * @param options
+   * @returns
+   */
+  toBinaryCoordinates(
+    options: {x: number; y: number; z: number} | ((data: number[], feature: {extent: any}) => void)
+  ): FlatFeature {
+    if (typeof options === 'function') {
+      return this._toBinaryCoordinates(options);
+    }
+    const tileIndex = options;
+    return this._toBinaryCoordinates((data: number[]) =>
+      projectToLngLatFlat(data, tileIndex, this.extent)
+    );
+  }
+
+  // eslint-disable-next-line max-statements
+  bbox() {
+    const pbf = this._pbf;
+    pbf.pos = this._geometry;
+
+    const end = pbf.readVarint() + pbf.pos;
+    let cmd = 1;
+    let length = 0;
+    let x = 0;
+    let y = 0;
+    let x1 = Infinity;
+    let x2 = -Infinity;
+    let y1 = Infinity;
+    let y2 = -Infinity;
+
+    while (pbf.pos < end) {
+      if (length <= 0) {
+        const cmdLen = pbf.readVarint();
+        cmd = cmdLen & 0x7;
+        length = cmdLen >> 3;
+      }
+
+      length--;
+
+      if (cmd === 1 || cmd === 2) {
+        x += pbf.readSVarint();
+        y += pbf.readSVarint();
+        if (x < x1) x1 = x;
+        if (x > x2) x2 = x;
+        if (y < y1) y1 = y;
+        if (y > y2) y2 = y;
+      } else if (cmd !== 7) {
+        throw new Error(`unknown command ${cmd}`);
+      }
+    }
+
+    return [x1, y1, x2, y2];
+  }
+
+  // BINARY HELPERS
+
+  /**
+   *
+   * @param transform
+   * @returns result
+   */
+  _toBinaryCoordinates(transform) {
+    // Expands the protobuf data to an intermediate Flat GeoJSON
+    // data format, which maps closely to the binary data buffers.
+    // It is similar to GeoJSON, but rather than storing the coordinates
+    // in multidimensional arrays, we have a 1D `data` with all the
+    // coordinates, and then index into this using the `indices`
+    // parameter, e.g.
+    //
+    // geometry: {
+    //   type: 'Point', data: [1,2], indices: [0]
+    // }
+    // geometry: {
+    //   type: 'LineString', data: [1,2,3,4,...], indices: [0]
+    // }
+    // geometry: {
+    //   type: 'Polygon', data: [1,2,3,4,...], indices: [[0, 2]]
+    // }
+    // Thus the indices member lets us look up the relevant range
+    // from the data array.
+    // The Multi* versions of the above types share the same data
+    // structure, just with multiple elements in the indices array
+    const geom = this.loadFlatGeometry();
+    let geometry;
+
+    // Apply the supplied transformation to data
+    transform(geom.data, this);
+
+    const coordLength = 2;
+
+    // eslint-disable-next-line default-case
+    switch (this.type) {
+      case 1: // Point
+        this._geometryInfo.pointFeaturesCount++;
+        this._geometryInfo.pointPositionsCount += geom.indices.length;
+        geometry = {type: 'Point', ...geom};
+        break;
+
+      case 2: // LineString
+        this._geometryInfo.lineFeaturesCount++;
+        this._geometryInfo.linePathsCount += geom.indices.length;
+        this._geometryInfo.linePositionsCount += geom.data.length / coordLength;
+        geometry = {type: 'LineString', ...geom};
+        break;
+
+      case 3: // Polygon
+        geometry = classifyRingsFlat(geom);
+
+        // Unlike Point & LineString geom.indices is a 2D array, thanks
+        // to the classifyRings method
+        this._geometryInfo.polygonFeaturesCount++;
+        this._geometryInfo.polygonObjectsCount += geometry.indices.length;
+
+        for (const indices of geometry.indices) {
+          this._geometryInfo.polygonRingsCount += indices.length;
+        }
+        this._geometryInfo.polygonPositionsCount += geometry.data.length / coordLength;
+
+        break;
+      default:
+        throw new Error(`Invalid geometry type: ${this.type}`);
+    }
+
+    const result: FlatFeature = {type: 'Feature', geometry, properties: this.properties};
+
+    if (this.id !== null) {
+      result.id = this.id;
+    }
+
+    return result;
+  }
+
+  // GEOJSON HELPER
 
   // eslint-disable-next-line complexity, max-statements
   loadGeometry(): number[][][] {
@@ -93,24 +259,30 @@ export class VectorTileFeature {
     return lines;
   }
 
-  // eslint-disable-next-line max-statements
-  bbox() {
+  // eslint-disable-next-line complexity, max-statements
+  loadFlatGeometry(): FlatIndexedGeometry {
     const pbf = this._pbf;
     pbf.pos = this._geometry;
 
-    const end = pbf.readVarint() + pbf.pos;
+    const endPos = pbf.readVarint() + pbf.pos;
     let cmd = 1;
+    let cmdLen: number;
     let length = 0;
     let x = 0;
     let y = 0;
-    let x1 = Infinity;
-    let x2 = -Infinity;
-    let y1 = Infinity;
-    let y2 = -Infinity;
+    let i = 0;
 
-    while (pbf.pos < end) {
+    // Note: I attempted to replace the `data` array with a
+    // Float32Array, but performance was worse, both using
+    // `set()` and direct index access. Also, we cannot
+    // know how large the buffer should be, so it would
+    // increase memory usage
+    const indices: number[] = []; // Indices where geometries start
+    const data: number[] = []; // Flat array of coordinate data
+
+    while (pbf.pos < endPos) {
       if (length <= 0) {
-        const cmdLen = pbf.readVarint();
+        cmdLen = pbf.readVarint();
         cmd = cmdLen & 0x7;
         length = cmdLen >> 3;
       }
@@ -120,40 +292,26 @@ export class VectorTileFeature {
       if (cmd === 1 || cmd === 2) {
         x += pbf.readSVarint();
         y += pbf.readSVarint();
-        if (x < x1) x1 = x;
-        if (x > x2) x2 = x;
-        if (y < y1) y1 = y;
-        if (y > y2) y2 = y;
-      } else if (cmd !== 7) {
+
+        if (cmd === 1) {
+          // New line
+          indices.push(i);
+        }
+        data.push(x, y);
+        i += 2;
+      } else if (cmd === 7) {
+        // Workaround for https://github.com/mapbox/mapnik-vector-tile/issues/90
+        if (i > 0) {
+          const start = indices[indices.length - 1]; // start index of polygon
+          data.push(data[start], data[start + 1]); // closePolygon
+          i += 2;
+        }
+      } else {
         throw new Error(`unknown command ${cmd}`);
       }
     }
 
-    return [x1, y1, x2, y2];
-  }
-
-  toGeoJSON(
-    options: {x: number; y: number; z: number} | ((data: number[], feature: {extent: any}) => void)
-  ): Feature {
-    const coords = this.loadGeometry();
-
-    if (typeof options === 'function') {
-      return _toGeoJSON(this, coords, options);
-    }
-    const {x, y, z} = options;
-    const size = this.extent * Math.pow(2, z);
-    const x0 = this.extent * x;
-    const y0 = this.extent * y;
-
-    function project(line: number[]) {
-      for (let j = 0; j < line.length; j++) {
-        const p = line[j];
-        p[0] = ((p[0] + x0) * 360) / size - 180;
-        const y2 = 180 - ((p[1] + y0) * 360) / size;
-        p[1] = (360 / Math.PI) * Math.atan(Math.exp((y2 * Math.PI) / 180)) - 90;
-      }
-    }
-    return _toGeoJSON(this, coords, project);
+    return {data, indices};
   }
 }
