+++ modules/mvt/src/lib/vector-tile/vector-tile-layer.ts
@@ -7,6 +7,7 @@
 
 import Protobuf from 'pbf';
 import {VectorTileFeature} from './vector-tile-feature';
+import {GeojsonGeometryInfo} from '@loaders.gl/schema';
 
 export class VectorTileLayer {
   version: number;
@@ -17,6 +18,7 @@ export class VectorTileLayer {
   _keys: string[];
   _values: (string | number | boolean | null)[];
   _features: number[];
+
   constructor(pbf: Protobuf, end: number) {
     // Public
     this.version = 1;
@@ -40,8 +42,7 @@ export class VectorTileLayer {
    * @param index
    * @returns feature
    */
-
-  feature(i: number): VectorTileFeature {
+  getGeoJSONFeature(i: number): VectorTileFeature {
     if (i < 0 || i >= this._features.length) {
       throw new Error('feature index out of bounds');
     }
@@ -50,6 +51,31 @@ export class VectorTileLayer {
 
     const end = this._pbf.readVarint() + this._pbf.pos;
     return new VectorTileFeature(this._pbf, end, this.extent, this._keys, this._values);
+  }
+
+  /**
+   * return binary feature `i` from this layer as a `VectorTileFeature`
+   *
+   * @param index
+   * @param geometryInfo
+   * @returns binary feature
+   */
+  getBinaryFeature(i: number, geometryInfo: GeojsonGeometryInfo): VectorTileFeature {
+    if (i < 0 || i >= this._features.length) {
+      throw new Error('feature index out of bounds');
+    }
+
+    this._pbf.pos = this._features[i];
+
+    const end = this._pbf.readVarint() + this._pbf.pos;
+    return new VectorTileFeature(
+      this._pbf,
+      end,
+      this.extent,
+      this._keys,
+      this._values,
+      geometryInfo
+    );
   }
 }
